Move extlib into a dedicated folder + license
