+++ Sofa/GUI/Batch/CMakeLists.txt
@@ -6,9 +6,9 @@ set(SOFAGUIBATCH_ROOT src/sofa/gui/batch)
 set(HEADER_FILES
     ${SOFAGUIBATCH_ROOT}/config.h.in
     ${SOFAGUIBATCH_ROOT}/init.h
-    ${SOFAGUIBATCH_ROOT}/indicators.hpp
     ${SOFAGUIBATCH_ROOT}/BatchGUI.h
     ${SOFAGUIBATCH_ROOT}/ProgressBar.h
+    ${SOFAGUIBATCH_ROOT}/indicators/indicators.hpp
 )
 
 set(SOURCE_FILES
